Completed module

+++ GradeBook/GradeBooks/BaseGradeBook.cs
@@ -107,7 +107,7 @@ namespace GradeBook.GradeBooks
             }
         }
 
-        public virtual int GetGPA(char letterGrade, StudentType studentType)
+        public virtual double GetGPA(char letterGrade, StudentType studentType)
         {
             var gpa = 0;
             switch (letterGrade)
